+++ test/fixtures/files/locked_import_map.rb
@@ -1,0 +1,3 @@
+pin "react" # @17.0.2 (locked)
+pin "md5", to: "https://cdn.jsdelivr.net/npm/md5@2.2.0/md5.js", preload: false # @2.2.0 (locked)
+pin "luxon" # @3.7.2 (esm.run, minified)